Added save functionality to library
Added a save_img function and take_and_save function to the library, and updated the examples to match. This puts all the saving right into the library, as it was the same across all platforms. Should work just fine with any existing code. But new code can just use a single function to take *and* save a picture!