Split to separate class
Split into a legacy (doesn't break code) class and a VC0706Camera class that is intended to be in line with the camera API. Some tweaks to make Pylint happy. Left in the save_image bit to the legacy version, so there's an easy way to save files there (Although take_and_save is gone).

Camera version doesn't use fileformat currently because I'm not sure how it should be checked. It's not super important since the camera only suports one.

I've had some issues changing between image resolutions on the camera, but that's likely something specifically with the camera.