Cleaning size changes
Both classes: Adds a reset() function, which resets the camera and resets the baudrate to whatever it was previously. This seems to be required in order to change the camera image size on the one I have, at least. This required adding an internal set_baudrate.

External: Added a get_image_size which converts width and height arguments into the proper const.

Added some retries for reading the buffer for VC0706Camera's take_picture.

Added the "Failed to halt frame!" to the stop_video instead of having it external.

Still need to handle that filetype.